Several Changes
- Deleted old project LibHKOSharp with files backed up in Temp_OldCode_dotNetFramework dir
- Removed project DebugTool fron solution
