+++ Temp_OldCode_dotNetFramework/LibHKOSharp/Properties/AssemblyInfo.cs
