+++ Temp_OldCode_dotNetFramework/LibHKOSharp/WeatherClass/LocalWeatherForecast.cs
