+++ Temp_OldCode_dotNetFramework/LibHKOSharp/WeatherClass/NineDaysWeather.cs
