+++ Temp_OldCode_dotNetFramework/LibHKOSharp/WeatherClass/Weather.cs
