+++ Temp_OldCode_dotNetFramework/LibHKOSharp/packages.config
